Enable test for sentence split
