--- RecDP/pyrecdp/pipeline/config/__init__.py
@@ -1,1 +1,0 @@
-from .config import *
