+++ RecDP/pyrecdp/pipeline/core/__init__.py
@@ -1,1 +1,2 @@
 from .pipeline import Pipeline
+from .config import *
