+++ RecDP/setup.py
@@ -76,7 +76,6 @@ setuptools.setup(
         "tabulate==0.9.0",
         "sentencepiece==0.1.99",
         "selectolax"
-        "loguru",
         "ray",
         "Faker"
         ])
